Extract MongoDB functionality into db.py
