+++ providers/ffvl.py
@@ -77,10 +77,9 @@ class Ffvl(Provider):
                         raise ProviderException(f"Unknown station '{station_id}'")
                     station = stations[station_id]
 
-                    measures_collection = self.measures_collection(station_id)
                     key = arrow.get(ffvl_measure["date"], "YYYY-MM-DD HH:mm:ss").replace(tzinfo=ffvl_tz).int_timestamp
 
-                    if not self.has_measure(measures_collection, key):
+                    if not self.has_measure(station_id, key):
                         measure = self.create_measure(
                             station,
                             key,
@@ -91,7 +90,7 @@ class Ffvl(Provider):
                             humidity=ffvl_measure["hydrometrie"],
                             pressure=Pressure(qfe=ffvl_measure["pression"], qnh=None, qff=None),
                         )
-                        self.insert_new_measures(measures_collection, station, [measure])
+                        self.insert_new_measures(station_id, station, [measure])
 
                 except ProviderException as e:
                     self.log.warning(f"Error while processing measures for station '{station_id}': {e}")
