+++ providers/meteoswiss_opendata.py
@@ -135,7 +135,6 @@ class MeteoSwiss(Provider):
                         meteoswiss_station["properties"]["reference_ts"], "YYYY-MM-DDTHH:mm:ssZ"
                     ).int_timestamp
 
-                    measures_collection = self.measures_collection(station_id)
                     new_measures = []
 
                     if meteoswiss_id in temperature_data:
@@ -165,7 +164,7 @@ class MeteoSwiss(Provider):
                     else:
                         rain = None
 
-                    if not self.has_measure(measures_collection, key):
+                    if not self.has_measure(station_id, key):
                         measure = self.create_measure(
                             station,
                             key,
@@ -179,7 +178,7 @@ class MeteoSwiss(Provider):
                         )
                         new_measures.append(measure)
 
-                    self.insert_new_measures(measures_collection, station, new_measures)
+                    self.insert_new_measures(station_id, station, new_measures)
 
                 except ProviderException as e:
                     self.log.warning(f"Error while processing station '{station_id}': {e}")
