+++ providers/pdcs.py
@@ -33,10 +33,9 @@ class Pdcs(Provider):
                         measure = pdcs_station["measurement"][0]
                         key = measure["time"]
 
-                        measures_collection = self.measures_collection(station_id)
                         new_measures = []
 
-                        if not self.has_measure(measures_collection, key):
+                        if not self.has_measure(station_id, key):
                             measure = self.create_measure(
                                 station,
                                 key,
@@ -46,7 +45,7 @@ class Pdcs(Provider):
                                 pressure=Pressure(qfe=measure["pres"]["qfe"], qnh=None, qff=None),
                             )
                             new_measures.append(measure)
-                            self.insert_new_measures(measures_collection, station, new_measures)
+                            self.insert_new_measures(station_id, station, new_measures)
 
                 except ProviderException as e:
                     self.log.warning(f"Error while processing station '{station_id}': {e}")
