+++ providers/windspots.py
@@ -67,8 +67,7 @@ class Windspots(Provider):
                                 f"inconsistent with measure time '{key_time}'"
                             )
 
-                        measures_collection = self.measures_collection(station_id)
-                        if not self.has_measure(measures_collection, key):
+                        if not self.has_measure(station_id, key):
                             try:
                                 measure = self.create_measure(
                                     station,
@@ -79,7 +78,7 @@ class Windspots(Provider):
                                     temperature=windspots_measure.get("airTemperature"),
                                     humidity=windspots_measure.get("airHumidity"),
                                 )
-                                self.insert_new_measures(measures_collection, station, [measure])
+                                self.insert_new_measures(station_id, station, [measure])
                             except ProviderException as e:
                                 self.log.warning(
                                     f"Error while processing measure '{key}' for station '{station_id}': {e}"
