+++ providers/zermatt.py
@@ -126,8 +126,7 @@ class Zermatt(Provider):
                                     .int_timestamp
                                 )
 
-                                measures_collection = self.measures_collection(station_id)
-                                if not self.has_measure(measures_collection, key):
+                                if not self.has_measure(station_id, key):
                                     wind_dir_text = stations[i + 1].xpath("td[@class='c4']")[0].text
                                     wind_dir = self.wind_directions[wind_dir_text.strip()]
 
@@ -144,7 +143,7 @@ class Zermatt(Provider):
                                     measure = self.create_measure(
                                         station, key, wind_dir, wind_avg, wind_avg, temperature=temp
                                     )
-                                    self.insert_new_measures(measures_collection, station, [measure])
+                                    self.insert_new_measures(station_id, station, [measure])
                             else:
                                 self.log.warning(f"No data for station '{station_id}'")
                         except ProviderException as e:
